Fix kubernetes provider synchronization issues (#802)
* Remove unused fn argument

fn: KubeClient::createResource
arg: waitReady

* Fix KubeClient::wait_transfer_container()

- We were assessing if the transfer-files-container is ready only by
checking if it's state is "running". While this is necessary, it is not
enough. We need to make sure that it finished initializing (downloading
coreutils). We do this by looking for the "waiting for tar to finish"
line in the logs
- Define a reusable method for retrying actions.

* Fix k8s provider setupChainSpec()

- Wait for the chain to start before trying to copy the chain spec file
- Use wait_log() for getChainSpecRaw()

* Fix wget failures

I stumbled upon a lot of failures when downloading coreutils, after curl
had been downloaded successfully from the same server. Not sure why.
Might be an issue with the wget used in the container. Anyway, not using
https seems to fix the issue.